🔥 refactor: clean code
